# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GenerateKeyBodyScope =
  (typeof GenerateKeyBodyScope)[keyof typeof GenerateKeyBodyScope];

export const GenerateKeyBodyScope = {
  read: "read",
  write: "write",
  admin: "admin",
} as const;

export interface GenerateKeyBody {
  name: string;
  scope: GenerateKeyBodyScope;
}

export interface ApiKeyResponse {
  key: string;
  name: string;
  scope: string;
  createdAt: string;
}

export type ProjectRuntime =
  (typeof ProjectRuntime)[keyof typeof ProjectRuntime];

export const ProjectRuntime = {
  static: "static",
  nodejs: "nodejs",
  python: "python",
  java: "java",
  unknown: "unknown",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  idle: "idle",
  building: "building",
  live: "live",
  failed: "failed",
  taken_down: "taken_down",
} as const;

export interface Project {
  id: string;
  name: string;
  description?: string;
  runtime: ProjectRuntime;
  status: ProjectStatus;
  liveUrl?: string;
  repoUrl?: string;
  createdAt: string;
  updatedAt: string;
  deploymentCount: number;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
  repoUrl?: string;
}

export type DeploymentStatus =
  (typeof DeploymentStatus)[keyof typeof DeploymentStatus];

export const DeploymentStatus = {
  queued: "queued",
  cloning: "cloning",
  detecting: "detecting",
  building: "building",
  deploying: "deploying",
  live: "live",
  failed: "failed",
  taken_down: "taken_down",
} as const;

export type DeploymentRuntime =
  (typeof DeploymentRuntime)[keyof typeof DeploymentRuntime];

export const DeploymentRuntime = {
  static: "static",
  nodejs: "nodejs",
  python: "python",
  java: "java",
  unknown: "unknown",
} as const;

export type DeploymentEnvVars = { [key: string]: string };

export interface Deployment {
  id: string;
  projectId: string;
  projectName: string;
  status: DeploymentStatus;
  runtime: DeploymentRuntime;
  branch?: string;
  commitHash?: string;
  liveUrl?: string;
  repoUrl?: string;
  envVars?: DeploymentEnvVars;
  startedAt?: string;
  completedAt?: string;
  createdAt: string;
}

export type TriggerDeploymentBodyEnvVars = { [key: string]: string };

export interface TriggerDeploymentBody {
  projectId: string;
  repoUrl?: string;
  branch?: string;
  envVars?: TriggerDeploymentBodyEnvVars;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  success: "success",
} as const;

export interface LogEntry {
  timestamp: string;
  level: LogEntryLevel;
  message: string;
}

export interface DeploymentLogs {
  deploymentId: string;
  logs: LogEntry[];
}

export interface AdminStats {
  totalProjects: number;
  totalDeployments: number;
  activeDeployments: number;
  failedDeployments: number;
  activeWebSocketConnections: number;
  serverLoad: number;
  uptime: number;
  masterApiKey?: string;
}

export type DashboardSummaryRuntimeBreakdown = {
  static: number;
  nodejs: number;
  python: number;
  java: number;
};

export interface DashboardSummary {
  totalProjects: number;
  liveDeployments: number;
  buildsToday: number;
  successRate: number;
  runtimeBreakdown: DashboardSummaryRuntimeBreakdown;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  deployed: "deployed",
  failed: "failed",
  taken_down: "taken_down",
  created: "created",
  deleted: "deleted",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  projectName: string;
  deploymentId?: string;
  message: string;
  timestamp: string;
}

export type ListDeploymentsParams = {
  projectId?: string;
};