# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate API Key
 */
export const GenerateApiKeyBody = zod.object({
  name: zod.string(),
  scope: zod.enum(["read", "write", "admin"]),
});

export const GenerateApiKeyResponse = zod.object({
  key: zod.string(),
  name: zod.string(),
  scope: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  runtime: zod.enum(["static", "nodejs", "python", "java", "unknown"]),
  status: zod.enum(["idle", "building", "live", "failed", "taken_down"]),
  liveUrl: zod.string().optional(),
  repoUrl: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  deploymentCount: zod.number(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  repoUrl: zod.string().optional(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.string(),
});

export const GetProjectResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  runtime: zod.enum(["static", "nodejs", "python", "java", "unknown"]),
  status: zod.enum(["idle", "building", "live", "failed", "taken_down"]),
  liveUrl: zod.string().optional(),
  repoUrl: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  deploymentCount: zod.number(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Trigger a new deployment
 */
export const TriggerDeploymentBody = zod.object({
  projectId: zod.string(),
  repoUrl: zod.string().optional(),
  branch: zod.string().optional(),
  envVars: zod.record(zod.string(), zod.string()).optional(),
});

/**
 * @summary List all deployments
 */
export const ListDeploymentsQueryParams = zod.object({
  projectId: zod.coerce.string().optional(),
});

export const ListDeploymentsResponseItem = zod.object({
  id: zod.string(),
  projectId: zod.string(),
  projectName: zod.string(),
  status: zod.enum([
    "queued",
    "cloning",
    "detecting",
    "building",
    "deploying",
    "live",
    "failed",
    "taken_down",
  ]),
  runtime: zod.enum(["static", "nodejs", "python", "java", "unknown"]),
  branch: zod.string().optional(),
  commitHash: zod.string().optional(),
  liveUrl: zod.string().optional(),
  repoUrl: zod.string().optional(),
  envVars: zod.record(zod.string(), zod.string()).optional(),
  startedAt: zod.string().optional(),
  completedAt: zod.string().optional(),
  createdAt: zod.string(),
});
export const ListDeploymentsResponse = zod.array(ListDeploymentsResponseItem);

/**
 * @summary Get deployment details
 */
export const GetDeploymentParams = zod.object({
  id: zod.coerce.string(),
});

export const GetDeploymentResponse = zod.object({
  id: zod.string(),
  projectId: zod.string(),
  projectName: zod.string(),
  status: zod.enum([
    "queued",
    "cloning",
    "detecting",
    "building",
    "deploying",
    "live",
    "failed",
    "taken_down",
  ]),
  runtime: zod.enum(["static", "nodejs", "python", "java", "unknown"]),
  branch: zod.string().optional(),
  commitHash: zod.string().optional(),
  liveUrl: zod.string().optional(),
  repoUrl: zod.string().optional(),
  envVars: zod.record(zod.string(), zod.string()).optional(),
  startedAt: zod.string().optional(),
  completedAt: zod.string().optional(),
  createdAt: zod.string(),
});

/**
 * @summary Takedown a deployment
 */
export const DeleteDeploymentParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get deployment logs
 */
export const GetDeploymentLogsParams = zod.object({
  deploymentId: zod.coerce.string(),
});

export const GetDeploymentLogsResponse = zod.object({
  deploymentId: zod.string(),
  logs: zod.array(
    zod.object({
      timestamp: zod.string(),
      level: zod.enum(["info", "warn", "error", "success"]),
      message: zod.string(),
    }),
  ),
});

/**
 * @summary Get system stats
 */
export const GetAdminStatsResponse = zod.object({
  totalProjects: zod.number(),
  totalDeployments: zod.number(),
  activeDeployments: zod.number(),
  failedDeployments: zod.number(),
  activeWebSocketConnections: zod.number(),
  serverLoad: zod.number(),
  uptime: zod.number(),
  masterApiKey: zod.string().optional(),
});

/**
 * @summary Get dashboard summary
 */
export const GetDashboardSummaryResponse = zod.object({
  totalProjects: zod.number(),
  liveDeployments: zod.number(),
  buildsToday: zod.number(),
  successRate: zod.number(),
  runtimeBreakdown: zod.object({
    static: zod.number(),
    nodejs: zod.number(),
    python: zod.number(),
    java: zod.number(),
  }),
});

/**
 * @summary Get recent activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.string(),
  type: zod.enum(["deployed", "failed", "taken_down", "created", "deleted"]),
  projectName: zod.string(),
  deploymentId: zod.string().optional(),
  message: zod.string(),
  timestamp: zod.string(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);