# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalProjects: number;
  totalDeployments: number;
  activeDeployments: number;
  failedDeployments: number;
  activeWebSocketConnections: number;
  serverLoad: number;
  uptime: number;
  masterApiKey?: string;
}