# lib/api-zod/src/generated/types/apiKeyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyResponse {
  key: string;
  name: string;
  scope: string;
  createdAt: string;
}