# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRuntimeBreakdown } from "./dashboardSummaryRuntimeBreakdown";

export interface DashboardSummary {
  totalProjects: number;
  liveDeployments: number;
  buildsToday: number;
  successRate: number;
  runtimeBreakdown: DashboardSummaryRuntimeBreakdown;
}