# lib/api-zod/src/generated/types/deploymentLogs.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntry } from "./logEntry";

export interface DeploymentLogs {
  deploymentId: string;
  logs: LogEntry[];
}