# lib/api-zod/src/generated/types/triggerDeploymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Hosting API - Instant deployment platform
 * OpenAPI spec version: 0.1.0
 */
import type { TriggerDeploymentBodyEnvVars } from "./triggerDeploymentBodyEnvVars";

export interface TriggerDeploymentBody {
  projectId: string;
  repoUrl?: string;
  branch?: string;
  envVars?: TriggerDeploymentBodyEnvVars;
}